Move guild converters to a separate helper class
